Move HTML parsing to own file
